Add devices code

See merge request plant-tracker/plant-tracker!24

+++ plant_listener/devices/include/plantlistener/device/device.hpp
@@ -1,8 +1,8 @@
 /*
- * (C) Copyright 2023 Brittany Mueller and Larry Milne (https://www.larrycloud.ca)
+ * (C) Copyright 2023-2024 Brittany Mueller and Larry Milne (https://www.larrycloud.ca)
  *
  * This code is distributed on "AS IS" BASIS,
- * WITHOUT WARRANTINES OR CONDITIONS OF ANY KIND.
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
@@ -57,7 +57,7 @@ class Device {
    * @param port The port to be read.
    * @returns value of sensor between [min_value, max_value], or -1 on error.
    */
-  virtual uint64_t readPort(const uint8_t port) = 0;
+  virtual double readPort(const uint8_t port) = 0;
 
   virtual nlohmann::json dump();
 };
