+++ plant_listener/devices/mock/mock_device.hpp
@@ -1,8 +1,8 @@
 /*
- * (C) Copyright 2023 Brittany Mueller and Larry Milne (https://www.larrycloud.ca)
+ * (C) Copyright 2023-2024 Brittany Mueller and Larry Milne (https://www.larrycloud.ca)
  *
  * This code is distributed on "AS IS" BASIS,
- * WITHOUT WARRANTINES OR CONDITIONS OF ANY KIND.
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
@@ -21,6 +21,6 @@ class MockDevice : public Device {
   uint64_t fake_value_ = 128;
 
   MockDevice(const std::string& name = "mock_device", const DeviceType type = DeviceType::ADC, const uint8_t ports = 8);
-  uint64_t readPort(const uint8_t port) override;
+  double readPort(const uint8_t port) override;
 };
 }  // namespace plantlistener::device
